Relocate memory script to its own dir
